Use the mosi_idle_low flag to configure controller MOSI idle level
Now that this is supported in the kernel and py-spidev will soon
add an attribute mosi_idle_low, use the new mode flag to make sure
that the datastream has the proper format for the LEDs. Previously
this required kernel source modifications.

To make users aware of missing support for mosi_idle_low in the
spidev module or the kernel, add warning messages.

Signed-off-by: Oualid Derouiche <oualid.derouiche@kontron.de>
Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>